Added Workspaces output, Changed planning to receive workspaces list

+++ provisioner.py
@@ -1,3 +1,4 @@
+from cmath import log
 import boto3
 import sys
 import os
@@ -31,25 +32,30 @@ def log_function(func):
 
     return log
 
+
 @log_function
 def generate_ssl_cert():
     KEY_FILE = "Terraform/SSL/private.pem"
     CERT_FILE = "Terraform/SSL/selfsigned.pem"
-    k = crypto.PKey()
-    k.generate_key(crypto.TYPE_RSA, 4096)
-    cert = crypto.X509()
-    cert.set_serial_number(1000)
-    cert.gmtime_adj_notBefore(0)
-    cert.gmtime_adj_notAfter(10*365*24*60*60)
-    cert.get_subject().CN = "Kandula"
-    cert.set_issuer(cert.get_subject())
-    cert.set_pubkey(k)
-    cert.sign(k, 'sha512')
-    with open(CERT_FILE, "w") as f:
-        f.write(crypto.dump_certificate(
-            crypto.FILETYPE_PEM, cert).decode("utf-8"))
-    with open(KEY_FILE, "w") as f:
-        f.write(crypto.dump_privatekey(crypto.FILETYPE_PEM, k).decode("utf-8"))
+    if os.path.exists(KEY_FILE) or os.path.exists(CERT_FILE):
+        pass
+    else:
+        k = crypto.PKey()
+        k.generate_key(crypto.TYPE_RSA, 4096)
+        cert = crypto.X509()
+        cert.set_serial_number(1000)
+        cert.gmtime_adj_notBefore(0)
+        cert.gmtime_adj_notAfter(10*365*24*60*60)
+        cert.get_subject().CN = "Kandula"
+        cert.set_issuer(cert.get_subject())
+        cert.set_pubkey(k)
+        cert.sign(k, 'sha512')
+        with open(CERT_FILE, "w") as f:
+            f.write(crypto.dump_certificate(
+                crypto.FILETYPE_PEM, cert).decode("utf-8"))
+        with open(KEY_FILE, "w") as f:
+            f.write(crypto.dump_privatekey(
+                crypto.FILETYPE_PEM, k).decode("utf-8"))
 
 
 def progress(filename, size, sent):
@@ -289,47 +295,46 @@ def wait_for_plan_status(session: Session, run_id, statuses):
                 return
 
 
-def get_workspace_outputs(session: Session, workspace_id):
-    response = session.get(
-        f"https://app.terraform.io/api/v2/workspaces/{workspace_id}/current-state-version")
-    if response.status_code != 200:
-        exit()
-    else:
-        response_data = response.json()
-        state_id = response_data['data']['id']
+@log_function
+def get_workspace_outputs(session: Session, workspace_list):
+    all_outputs = {}
+    for workspace in workspace_list:
+        workspace_outputs_dict = {}
         response = session.get(
-            f"https://app.terraform.io/api/v2/state-versions/{state_id}/outputs")
+            f"https://app.terraform.io/api/v2/workspaces/{workspace['id']}")
         if response.status_code != 200:
             exit()
         else:
             response_data = response.json()
-            outputs_dict = {}
-            for output in response_data['data']:
-                output_key = output['data']['attributes']['name']
-                output_value = output['data']['attributes']['value']
-                outputs_dict[output_key] = output_value
-        return outputs_dict
+            output_ids = []
+            for output in response_data['data']['relationships']['outputs']['data']:
+                output_ids.append(output['id'])
+            for output_id in output_ids:
+                response = session.get(
+                    f"https://app.terraform.io/api/v2/state-version-outputs/{output_id}")
+                if response.status_code != 200:
+                    exit()
+                else:
+                    response_data = response.json()
+                    dict_key = response_data['data']['attributes']['name']
+                    dict_value = response_data['data']['attributes']['value']
+                    workspace_outputs_dict[dict_key] = dict_value
+        all_outputs[workspace['name']] = workspace_outputs_dict
+    return all_outputs
 
 
 @log_function
 # TODO Change from vpc_worspace to global value
-def run_plan_to_completion(session, organization_name, workspace_name):
-    vpc_workspace = get_tfe_workspace_by_names(
-        session,
-        organization_name,
-        workspace_name,
-    )
+def run_plan_to_completion(session, workspaces):
+    for workspace in workspaces:
+        current_plan = plan_by_workspace(session, workspace["id"])
+        current_plan_id = current_plan["data"]["id"]
 
-    vpc_plan = plan_by_workspace(session, vpc_workspace["id"])
-
-    vpc_plan_id = vpc_plan["data"]["id"]
-
-    wait_for_plan_status(session, vpc_plan_id, [
-                         "planned", "planned_and_finished"])
-    apply_run_by_id(session, vpc_plan_id)
-    wait_for_plan_status(session, vpc_plan_id, [
-                         "applied", "planned_and_finished"])
-    # outputs = get_workspace_outputs(session, vpc_workspace["id"])
+        wait_for_plan_status(session, current_plan_id, [
+            "planned", "planned_and_finished"])
+        apply_run_by_id(session, current_plan_id)
+        wait_for_plan_status(session, current_plan_id, [
+            "applied", "planned_and_finished"])
 
 
 ############################################################################################
@@ -346,23 +351,18 @@ def run_terraform(terraform_var_file_path, terraform_txt_vars):
     terraform_vars.update(terraform_txt_vars)
     terraform_vars.update(terraform_cloud_vars)
 
-    run_plan_to_completion(
-        session,
-        terraform_vars["tfe_organization_name"],
-        terraform_vars["vpc_workspace_name"],
-    )
+    vpc_workspace = get_tfe_workspace_by_names(
+        session, terraform_vars["tfe_organization_name"], terraform_vars["vpc_workspace_name"])
+    servers_workspace = get_tfe_workspace_by_names(
+        session, terraform_vars["tfe_organization_name"], terraform_vars["servers_workspace_name"])
+    kubernetes_workspace = get_tfe_workspace_by_names(
+        session, terraform_vars["tfe_organization_name"], terraform_vars["kubernetes_workspace_name"])
 
     run_plan_to_completion(
-        session,
-        terraform_vars["tfe_organization_name"],
-        terraform_vars["servers_workspace_name"],
+        session, [vpc_workspace, servers_workspace, kubernetes_workspace]
     )
 
-    run_plan_to_completion(
-        session,
-        terraform_vars["tfe_organization_name"],
-        terraform_vars["kubernetes_workspace_name"],
-    )
+    return get_workspace_outputs(session, [vpc_workspace, servers_workspace, kubernetes_workspace])
 
 
 @log_function
@@ -427,8 +427,9 @@ def run(vars_file_path):
     terraform_var_file_path = create_tfvars_file(vars_file_path)
     terraform_vars = get_terraform_vars_from_file(terraform_var_file_path)
     generate_ssl_cert()
-    run_terraform(terraform_var_file_path, terraform_vars)
+    workspaces_outputs = run_terraform(terraform_var_file_path, terraform_vars)
     run_ansible(terraform_vars)
+    print(json.dumps(workspaces_outputs, sort_keys=False, indent=4))
 
 
 # vars_file_path = sys.argv[1]
